@@ -1,7 +1,8 @@
 ﻿using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
+using RayWrapper.Objs;
 using RayWrapper.Vars;
-using static RayWrapper.FontManager;
+using static Raylib_CsLo.Raylib;
 
 namespace RayWrapper.Animation.SinglePurposeObjects
 {
@@ -10,6 +11,8 @@ namespace RayWrapper.Animation.SinglePurposeObjects
     /// </summary>
     public class MaskText : GameObject, ISizeable, IAlphable
     {
+        public static Text.Style defaultStyle = new();
+
         public override Vector2 Position
         {
             get => rect.Pos();
@@ -24,15 +27,27 @@ namespace RayWrapper.Animation.SinglePurposeObjects
             set => (_color, alpha) = (value, ((Color) value).a);
         }
 
+        public Text.Style style = defaultStyle.Copy();
         public int alpha = 255;
         public Actionable<string> text;
         public Rectangle rect;
-        public Actionable<string> tooltip = null;
+        public Tooltip tooltip;
 
-        private ColorModule _color = Color.WHITE;
+        private ColorModule _color = WHITE;
 
-        public MaskText(Actionable<string> text, Rectangle rect) => (this.text, this.rect) = (text, rect);
-        public MaskText(Rectangle rect) => this.rect = rect;
+        public MaskText(Rectangle rect) : this("", rect)
+        {
+        }
+
+        public MaskText(Actionable<string> text, Rectangle rect)
+        {
+            (this.text, this.rect) = (text, rect);
+            style.color = new ColorModule(() =>
+            {
+                var color = (Color) ColorMod;
+                return color.a != alpha ? color.SetAlpha(alpha) : color;
+            });
+        }
 
         protected override void UpdateCall()
         {
@@ -40,10 +55,8 @@ namespace RayWrapper.Animation.SinglePurposeObjects
 
         protected override void RenderCall()
         {
-            var color = (Color) ColorMod;
-            rect.Grow(2).MaskDraw(() =>
-                GetDefFont().DrawText(text, rect.Pos(), color.a != alpha ? color.SetAlpha(alpha) : color));
-            if (tooltip is not null) rect.DrawTooltip(tooltip);
+            rect.Grow(2).MaskDraw(() => style.Draw(text, rect));
+            tooltip?.Draw(rect);
         }
 
         public void SetSize(Vector2 size) => rect = rect.SetSize(size);
