@@ -1,5 +1,5 @@
 ﻿using System.Numerics;
-using RayWrapper.Vars;
+using RayWrapper.Var_Interfaces;
 
 namespace RayWrapper.Animation.Transitions
 {
