@@ -1,15 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Numerics;
-using Raylib_cs;
-using static Raylib_cs.Raylib;
+using Raylib_CsLo;
+using static Raylib_CsLo.Raylib;
 
 namespace RayWrapper.Objs
 {
     public class ImageMatrix
     {
         private Image _image;
-        private Dictionary<Enum, ImageObj> _slices = new();
+        private IDictionary<Enum, ImageObj> _slices = new Dictionary<Enum, ImageObj>();
 
         public ImageMatrix(string image) : this(LoadImage(image))
         {
