@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
+using RayWrapper.Var_Interfaces;
 using RayWrapper.Vars;
-using static Raylib_cs.Raylib;
-using static RayWrapper.FontManager;
+using static Raylib_CsLo.Raylib;
 using static RayWrapper.GameBox;
 using static RayWrapper.RectWrapper;
 
@@ -11,133 +11,148 @@ namespace RayWrapper.Objs
 {
     public class Label : GameObject
     {
-        public enum TextMode
-        {
-            AlignLeft,
-            AlignCenter,
-            AlignRight,
-            SizeToText,
-            WrapText
-        }
+        public static Style defaultStyle = new();
 
-        public Rectangle Rect => _rect;
+        public Rectangle Rect =>
+            style.drawMode is Style.DrawMode.SizeToText
+                ? AssembleRectFromVec(_back.Pos(), style.textStyle.MeasureText(text)).Grow(4)
+                : _back;
 
         public override Vector2 Position
         {
-            get => _back.Pos();
-            set
-            {
-                _back.MoveTo(value);
-                ReCalc();
-            }
+            get => Rect.Pos();
+            set => _back.MoveTo(value);
         }
 
-        public override Vector2 Size =>
-            textMode == TextMode.SizeToText ? _textSizeCache + new Vector2(8) : _back.Size();
+        public override Vector2 Size => Rect.Size();
 
-        public ColorModule backColor = new(50);
-        public ColorModule fontColor = new(192);
-        public ColorModule outlineColor = new(Color.BLACK);
+        public Style style = defaultStyle.Copy();
         public Action clicked;
-        public int fontSize = 24;
-        public bool outline = false;
-        public float spacing = 1.5f;
         public Actionable<string> text;
-        public Actionable<TextMode> textMode;
-        public Actionable<string> tooltip;
-        public Actionable<bool> useBaseHover = false;
+        public Tooltip tooltip;
 
         private Rectangle _back;
-        private string _textCache;
-        private Vector2 _textSizeCache;
 
-        // textmode cache
-        private Rectangle _rect;
-        private Rectangle _sizedRect;
-        private Rectangle _adj;
-        private Vector2 _sizedPos;
-        private Vector2 _alignLeft;
-        private Vector2 _alignCenter;
-        private Vector2 _alignRight;
-
-        public Label(Rectangle back, string text = "Untitled Label", TextMode textMode = TextMode.AlignLeft)
+        public Label(Vector2 pos, string text = "Untitled Label") : this(
+            AssembleRectFromVec(pos, defaultStyle.textStyle.MeasureText(text)).Grow(4), text)
         {
-            (_back, this.text, this.textMode) = (back, text, textMode);
         }
 
-        public Label(Vector2 pos, string text = "Untitled Label", TextMode textMode = TextMode.AlignLeft)
-        {
-            (this.text, this.textMode) = (text, textMode);
-            CheckText();
-            _back = AssembleRectFromVec(pos, _textSizeCache).Grow(4);
-        }
+        public Label(Rectangle back, string text = "Untitled Label") => (_back, this.text) = (back, text);
 
         protected override void UpdateCall()
         {
-            if (!Rect.IsMouseIn() || !IsMouseButtonPressed(MouseButton.MOUSE_LEFT_BUTTON)) return;
-            clicked?.Invoke();
+            if (Rect.IsMouseIn() && IsMouseButtonPressed(MOUSE_LEFT_BUTTON)) clicked?.Invoke();
         }
 
         protected override void RenderCall()
         {
-            var hover = useBaseHover && Rect.IsMouseIn() && !IsMouseOccupied;
-            Color realFc = hover ? ((Color) fontColor).MakeLighter() : fontColor;
-            Color realBc = hover ? ((Color) backColor).MakeLighter() : backColor;
-            CheckText();
-
-            void DrawTxt(Vector2 pos) => DrawTextEx(GetDefFont(fontSize), _textCache, pos, fontSize, spacing, realFc);
-
-            void DrawBack(Rectangle rect)
-            {
-                rect.DrawRounded(realBc, .25f);
-                if (outline) rect.DrawRoundedLines(outlineColor, .25f);
-            }
-
-            if (textMode != TextMode.SizeToText) DrawBack(_back);
-
-            switch ((TextMode) textMode)
-            {
-                case TextMode.AlignLeft:
-                    DrawTxt(_alignLeft);
-                    break;
-                case TextMode.AlignCenter:
-                    GetDefFont(fontSize).DrawCenterText(_alignCenter, _textCache, realFc, fontSize, spacing);
-                    break;
-                case TextMode.AlignRight:
-                    DrawTxt(_alignRight);
-                    break;
-                case TextMode.WrapText:
-                    GetDefFont(fontSize).DrawTextRec(_textCache, _adj, realFc, fontSize, spacing, true);
-                    break;
-                case TextMode.SizeToText:
-                    DrawBack(_sizedRect);
-                    DrawTxt(_sizedPos);
-                    break;
-                default:
-                    throw new ArgumentOutOfRangeException();
-            }
-
-            var t = (string) (tooltip ?? string.Empty);
-            if (Rect.IsMouseIn() && t != string.Empty && !IsMouseOccupied) Rect.DrawTooltip(t);
+            style.Draw(text, Rect);
+            if (tooltip is not null && !IsMouseOccupied) tooltip.Draw(Rect);
         }
 
-        public void CheckText()
+        public class Style : IStyle<Style>
         {
-            if (_textCache is not null && _textCache == text) return;
-            _textSizeCache = MeasureTextEx(GetDefFont(fontSize), _textCache = text, fontSize, spacing);
-            ReCalc();
-        }
+            public ColorModule fontColor = new(192);
+            public ColorModule backColor = new(70);
+            public Text.Style textStyle = new();
+            public RectStyle backStyle = new();
+            public OutlineStyle outline = new();
+            public bool drawOutline = true;
+            public Actionable<bool> drawHover = false;
+            public Func<Color, bool, Color> drawColor;
 
-        public void ReCalc()
-        {
-            _adj = _back.Shrink(4);
-            _sizedRect = AssembleRectFromVec(_adj.Pos(), _textSizeCache).Grow(4);
-            _sizedPos = _adj.Pos();
-            _alignLeft = new Vector2(_adj.x, _adj.y + (_adj.height / 2 - _textSizeCache.Y / 2));
-            _alignCenter = _adj.Center();
-            _alignRight = new Vector2(_adj.x + _adj.width - _textSizeCache.X,
-                _adj.y + (_adj.height / 2 - _textSizeCache.Y / 2));
-            _rect = AssembleRectFromVec(_back.Pos(), Size);
+            public DrawMode drawMode
+            {
+                get => _drawMode;
+                set
+                {
+                    textStyle.drawMode = value switch
+                    {
+                        DrawMode.AlignLeft => Text.Style.DrawMode.Normal,
+                        DrawMode.AlignCenter => Text.Style.DrawMode.Center,
+                        DrawMode.AlignRight => Text.Style.DrawMode.Normal,
+                        DrawMode.SizeToText => Text.Style.DrawMode.Normal,
+                        DrawMode.WrapText => Text.Style.DrawMode.Wrap,
+                        _ => throw new ArgumentOutOfRangeException(nameof(value), value, null)
+                    };
+                    _drawMode = value;
+                }
+            }
+
+            private DrawMode _drawMode = DrawMode.AlignLeft;
+            private bool hasOgDrawColor;
+
+            public enum DrawMode
+            {
+                AlignLeft,
+                AlignCenter,
+                AlignRight,
+                SizeToText,
+                WrapText
+            }
+
+            public Style(Func<Color, bool, Color> drawColor = null)
+            {
+                if (drawColor is not null)
+                {
+                    this.drawColor = drawColor;
+                    return;
+                }
+
+                this.drawColor = (c, b) =>
+                {
+                    if (!drawHover) return c;
+                    return b ? c.MakeLighter() : c;
+                };
+                hasOgDrawColor = true;
+            }
+
+            public void Draw(string text, Rectangle back)
+            {
+                var textSize = textStyle.MeasureText(text);
+                var shrink = back.Shrink(4);
+                backStyle.Draw(back);
+                if (drawOutline) outline.Draw(back);
+
+                var hover = back.IsMouseIn();
+                backStyle.color = drawColor.Invoke(backColor, hover);
+                textStyle.color = drawColor.Invoke(fontColor, hover);
+
+                switch (drawMode)
+                {
+                    case DrawMode.AlignLeft:
+                        textStyle.Draw(text,
+                            new Vector2(shrink.x, shrink.y + (shrink.height / 2 - textSize.Y / 2)));
+                        break;
+                    case DrawMode.AlignCenter:
+                        textStyle.Draw(text, shrink.Center());
+                        break;
+                    case DrawMode.AlignRight:
+                        textStyle.Draw(text,
+                            new Vector2(shrink.x + shrink.width - textSize.X,
+                                shrink.y + (shrink.height / 2 - textSize.Y / 2)));
+                        break;
+                    case DrawMode.SizeToText:
+                        textStyle.Draw(text, shrink.Pos());
+                        break;
+                    case DrawMode.WrapText:
+                        textStyle.Draw(text, shrink);
+                        break;
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
+            }
+
+            public Style Copy()
+            {
+                return new Style(hasOgDrawColor ? null : drawColor)
+                {
+                    outline = outline.Copy(), backColor = backColor.Copy(), backStyle = backStyle.Copy(),
+                    drawMode = drawMode, fontColor = fontColor.Copy(), textStyle = textStyle.Copy(),
+                    drawOutline = drawOutline, drawHover = drawHover
+                };
+            }
         }
     }
 }
