@@ -2,8 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
+using RayWrapper.Var_Interfaces;
 using RayWrapper.Vars;
+using ZimonIsHimUtils.ExtensionMethods;
+using static Raylib_CsLo.Raylib;
 
 namespace RayWrapper.Objs
 {
@@ -31,7 +34,7 @@ namespace RayWrapper.Objs
         private Vector2 _pos;
         private Vector2 _posOffset = Vector2.Zero;
         private Vector2 _trueSize;
-        
+
         public ScrollView(Rectangle rect)
         {
             _rect = new Rectangle(rect.x + 20, rect.y, rect.width - 20, rect.height - 20);
@@ -60,12 +63,12 @@ namespace RayWrapper.Objs
 
             _rect.MaskDraw(() =>
             {
-                foreach (var obj in _renderList) obj.Render();
+                _renderList.Each(obj => obj.Render());
             });
 
             if (_trueSize.X >= _size.X) _xScroll.Render();
             if (_trueSize.Y >= _size.Y) _yScroll.Render();
-            _rect.DrawHallowRect(Color.BLACK, 1);
+            _rect.DrawHallowRect(BLACK, 1);
         }
 
         public void Recalc()
@@ -80,7 +83,7 @@ namespace RayWrapper.Objs
             _posOffset = new Vector2(_xScroll.Value, _yScroll.Value) - new Vector2(3);
             var tempRect = new Rectangle(_rect.x + _posOffset.X, _rect.y + _posOffset.Y, _rect.width, _rect.height);
             _renderList.Clear();
-            _renderList = _gos.Where(g => Raylib.CheckCollisionRecs(g.GetDebugRect(), tempRect)).ToList();
+            _renderList = _gos.Where(g => CheckCollisionRecs(g.GetDebugRect(), tempRect)).ToList();
             foreach (var go in _renderList) go.Position -= _posOffset;
         }
 
