@@ -1,23 +1,25 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
+using RayWrapper.Var_Interfaces;
 using RayWrapper.Vars;
+using ZimonIsHimUtils.ExtensionMethods;
+using static Raylib_CsLo.Raylib;
 using static RayWrapper.GameBox;
 
 namespace RayWrapper.Objs
 {
     public class Scrollbar : GameObject
     {
-        public Func<float> amountInvoke;
+        public static Style defaultStyle = new();
+
+        public Style style = defaultStyle.Copy();
         public Rectangle bar = RectWrapper.Zero;
-        public ColorModule barColor = new(116, 116, 116);
-        public int minSizePercent = 20;
         public Rectangle container;
-        public ColorModule containerColor = new(78, 78, 78);
+        public int minSizePercent = 20;
         public bool isVertical;
-        public bool outline = true;
-        public ColorModule outlineColor = new(Color.BLACK);
+        public Func<float> amountInvoke;
 
         private readonly IList<Action<float>> _onMove = new List<Action<float>>();
         private float _trueSize;
@@ -60,7 +62,7 @@ namespace RayWrapper.Objs
             if (isVertical) bar.y -= offset;
             else bar.x -= offset;
             ClampBounds();
-            foreach (var a in _onMove) a.Invoke(Value);
+            _onMove.Each(a => a.Invoke(Value));
         }
 
         public void ClampBounds()
@@ -96,7 +98,7 @@ namespace RayWrapper.Objs
             ClampBounds();
             CalcVal();
 
-            var isLeft = Raylib.IsMouseButtonDown(MouseButton.MOUSE_LEFT_BUTTON);
+            var isLeft = IsMouseButtonDown(MOUSE_LEFT_BUTTON);
 
             if (bar.IsMouseIn() && isLeft && !IsMouseOccupied)
             {
@@ -107,7 +109,7 @@ namespace RayWrapper.Objs
 
             if (mouseOccupier != this)
             {
-                if (!IsMouseOccupied && Raylib.IsMouseButtonPressed(MouseButton.MOUSE_LEFT_BUTTON) &&
+                if (!IsMouseOccupied && IsMouseButtonPressed(MOUSE_LEFT_BUTTON) &&
                     container.IsMouseIn())
                     MoveBar((isVertical ? bar.y - mousePos.Y : bar.x - mousePos.X) + _visibleSize / 2);
 
@@ -122,14 +124,42 @@ namespace RayWrapper.Objs
         protected override void RenderCall()
         {
             if (Amount() == 1) return; // ignore loss of precision
-            var hover = IsMouseOccupied && mouseOccupier == this || !IsMouseOccupied && container.IsMouseIn();
-            container.DrawRounded(hover ? ((Color) containerColor).MakeLighter() : containerColor, .4f);
-            bar.DrawRounded(hover ? ((Color) barColor).MakeLighter() : barColor, .4f);
-            if (!outline) return;
-            container.DrawRoundedLines(outlineColor, .4f);
-            bar.DrawRoundedLines(outlineColor, .4f);
+            style.Draw(container, bar,
+                IsMouseOccupied && mouseOccupier == this || !IsMouseOccupied && container.IsMouseIn());
         }
 
         public float Amount() => Math.Max(amountInvoke?.Invoke() ?? 0, 1);
+
+        public class Style : IStyle<Style>
+        {
+            public RectStyle containerStyle = new();
+            public RectStyle barStyle = new();
+            public OutlineStyle containerOutline = new();
+            public OutlineStyle barOutline = new();
+            public ColorModule containerColor = new(78, 78, 78);
+            public ColorModule barColor = new(116, 116, 116);
+            public bool drawOutline = true;
+
+            public void Draw(Rectangle container, Rectangle bar, bool hover)
+            {
+                containerStyle.color = hover ? ((Color) containerColor).MakeLighter() : (Color) containerColor;
+                barStyle.color = hover ? ((Color) barColor).MakeLighter() : (Color) barColor;
+                containerStyle.Draw(container);
+                containerOutline.Draw(container);
+                if (!drawOutline) return;
+                barStyle.Draw(bar);
+                barOutline.Draw(bar);
+            }
+
+            public Style Copy()
+            {
+                return new Style
+                {
+                    containerStyle = containerStyle.Copy(), containerOutline = containerOutline.Copy(),
+                    containerColor = containerColor.Copy(), barStyle = barStyle.Copy(), barOutline = barOutline.Copy(),
+                    barColor = barColor.Copy(), drawOutline = drawOutline
+                };
+            }
+        }
     }
 }
