@@ -1,15 +1,15 @@
-﻿using System;
-using System.Numerics;
+﻿using System.Numerics;
+using RayWrapper.Vars;
 using static RayWrapper.RectWrapper;
 
 namespace RayWrapper.Objs.TreeView.TreeNodeChain.NodeShapes
 {
     public class Box : NodeShape
     {
-        public Box(Vector2 position, Vector2 size, Func<string> tooltip = null) : base(position, size) =>
+        public Box(Vector2 position, Vector2 size, Actionable<string> tooltip = null) : base(position, size) =>
             this.tooltip = tooltip;
 
-        public Box(Vector2 position, Func<string> tooltip = null) : base(position, Vector2.One) =>
+        public Box(Vector2 position, Actionable<string> tooltip = null) : base(position, Vector2.One) =>
             this.tooltip = tooltip;
 
         public override void DrawShape(Vector2 off, float scale) =>
