@@ -2,46 +2,73 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
 using RayWrapper.Objs.TreeView.TreeNodeChain;
 using RayWrapper.Vars;
-using static Raylib_cs.MouseButton;
-using static Raylib_cs.Raylib;
+using static Raylib_CsLo.Raylib;
 using static RayWrapper.GameBox;
 using static RayWrapper.RectWrapper;
 
 namespace RayWrapper.Objs.TreeView
 {
+    // todo: VVVVVV 
+    // - tree view: image node
     public class TreeView : GameObject
     {
+        public readonly List<NodeChain> chains = new();
+
         public override Vector2 Position { get; set; }
         public override Vector2 Size => mask.IsEqualTo(Zero) ? WindowSize : mask.Size();
 
+        public bool enableScaling = true;
+        public bool verticalMovement = true;
+        public bool horizontalMovement = true;
+        public float defaultScale = 32;
+        public float distanceThreshold = .15f;
+        public string selected;
+        public Rectangle bounds = Max;
         public Rectangle mask = Zero;
         public Vector2 axisOffset = Vector2.Zero;
-        public readonly List<NodeChain> chains = new();
+        public Tooltip tooltip;
 
         private Vector2 _lastPos;
         private Vector2 _moveChange;
-        private float _scale = 32;
+        private float _scale;
+        private List<string> _tooltipList = new();
 
-        public TreeView(params NodeChain[] chains) => this.chains.AddRange(chains);
+        public TreeView(params NodeChain[] chains)
+        {
+            _scale = defaultScale;
+            this.chains.AddRange(chains);
+            tooltip = new DefaultTooltip(new Actionable<string>(() => selected));
+        }
 
         protected override void UpdateCall()
         {
-            if (alertBox is not null || IsMouseOccupied && mouseOccupier != this) return;
+            if (alertQueue.Count > 0 || IsMouseOccupied && mouseOccupier != this) return;
             if (IsMouseButtonPressed(MOUSE_MIDDLE_BUTTON)) ResetPos();
             if (IsMouseButtonDown(MOUSE_LEFT_BUTTON) && (mask.IsEqualTo(Zero)
-                ? AssembleRectFromVec(Vector2.Zero, WindowSize)
-                : mask).IsMouseIn())
+                    ? AssembleRectFromVec(Vector2.Zero, WindowSize)
+                    : mask).IsMouseIn())
             {
                 var curMouse = mousePos;
                 if (mouseOccupier != this) _lastPos = curMouse;
-                _moveChange += (curMouse - _lastPos) / (_scale / 1.4f);
+                var toMove = (curMouse - _lastPos) / (_scale / 1.4f);
+
+                if (Vector2.Zero.Distance(toMove) <= distanceThreshold)
+                {
+                    (_lastPos, mouseOccupier) = (curMouse, this);
+                    return;
+                }
+
+                _moveChange += new Vector2(horizontalMovement ? toMove.X : 0, verticalMovement ? toMove.Y : 0);
+                _moveChange = new Vector2(Math.Clamp(_moveChange.X, bounds.x, bounds.width),
+                    Math.Clamp(_moveChange.Y, bounds.y, bounds.height));
                 (_lastPos, mouseOccupier) = (curMouse, this);
             }
             else mouseOccupier = null;
 
+            if (!enableScaling) return;
             var scroll = GetMouseWheelMove();
             if (scroll != 0) _scale = Math.Min(Math.Max(15, _scale + scroll), 55);
         }
@@ -53,10 +80,17 @@ namespace RayWrapper.Objs.TreeView
                 ? AssembleRectFromVec(Vector2.Zero, WindowSize)
                 : mask).MaskDraw(() =>
             {
-                foreach (var nodeChain in chains) nodeChain.Draw((_moveChange + axisOffset) * _scale, _scale);
+                _tooltipList.AddRange(chains
+                    .Select(nodeChain => nodeChain.Draw((_moveChange + axisOffset) * _scale, _scale)));
             });
+
+            var remain = _tooltipList.Where(i => i is not null).ToArray();
+            _tooltipList.Clear();
+            if (!remain.Any()) return;
+            selected = remain.First();
+            tooltip.Draw();
         }
 
-        public void ResetPos() => (_moveChange, _scale) = (Vector2.Zero, _scale = 32);
+        public void ResetPos() => (_moveChange, _scale) = (Vector2.Zero, _scale = defaultScale);
     }
 }
