@@ -1,7 +1,7 @@
 ﻿using System.Numerics;
-using Raylib_cs;
 using RayWrapper;
-using RayWrapper.CollisionSystem;
+using RayWrapper.Collision;
+using static Raylib_CsLo.Raylib;
 
 namespace RayWrapperTesterCollision
 {
@@ -9,11 +9,12 @@ namespace RayWrapperTesterCollision
     {
         public bool vert;
 
-        public Bar(Vector2 pos, Vector2 size) : base("bar", pos, new Rectangle(0, 0, size.X, size.Y))
+        public Bar(Vector2 pos, Vector2 size) : base(RectWrapper.AssembleRectFromVec(pos, size))
         {
+            tag = "bar";
         }
 
-        public override void RenderShape(Vector2 pos) => rect.NewMoveTo(pos).Draw(Color.GRAY);
+        public override void RenderShape(Vector2 pos) => rect.NewMoveTo(pos).Draw(GRAY);
 
         public override void FirstCollision(Collider c)
         {
