@@ -1,11 +1,10 @@
 ﻿using System.Numerics;
-using Raylib_cs;
+using Raylib_CsLo;
 using RayWrapper;
-using RayWrapper.CollisionSystem;
 using RayWrapper.Objs;
 using RayWrapper.Vars;
-using static Raylib_cs.Color;
-using static Raylib_cs.Raylib;
+using static Raylib_CsLo.Raylib;
+using static RayWrapper.Collision.Collision;
 using static RayWrapper.GameBox;
 
 namespace RayWrapperTesterCollision
@@ -19,16 +18,19 @@ namespace RayWrapperTesterCollision
 
         public override void Init()
         {
-            InitCollision();
+            InitPhysics(8, 8);
             var wx = WindowSize.X;
             var wy = WindowSize.Y;
-            CollisionLayerTags.Add(("circle", "bar"));
+
+            collisionRules.TryAdd("bar", "ball");
+
             new Bar(new Vector2(0, 0), new Vector2(wx, 10));
             new Bar(new Vector2(0, wy - 10), new Vector2(wx, 10));
             new Bar(new Vector2(0, 0), new Vector2(10, wy)) { vert = true };
             new Bar(new Vector2(wx - 10, 0), new Vector2(10, wy)) { vert = true };
-            RegisterGameObj(new Text($"{Collider.count - 4}", new Vector2(12, 60), RED), new Text(
-                new Actionable<string>(() => $@"Collision Time
+            RegisterGameObj(new Text(new Actionable<string>(() => $"{CountColliders() - 4}"), new Vector2(12, 60), RED),
+                new Text(
+                    new Actionable<string>(() => $@"Collision Time
 cur: {CurrentCollision}ms
 avg: {TimeAverage}ms
 high: {CollisionHigh}ms".Replace("\r", "")), new Vector2(300, 50), SKYBLUE));
@@ -36,8 +38,10 @@ high: {CollisionHigh}ms".Replace("\r", "")), new Vector2(300, 50), SKYBLUE));
 
         public override void UpdateLoop()
         {
-            if (!IsMouseButtonPressed(MouseButton.MOUSE_LEFT_BUTTON)) return;
-            for (var i = 0; i < 50; i++) new Circle(mousePos);
+            if (IsMouseButtonPressed(MouseButton.MOUSE_BUTTON_LEFT))
+            {
+                for (var i = 0; i < 100; i++) new Circle(mousePos);
+            }
         }
 
         public override void RenderLoop() => DrawFPS(12, 12);
